+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Changes to procedures and constraints:
   * Merged `Another layer to align layers with` into `Object to align layers with`.
   * Units for x- and y-offsets are moved next to the x- and y-offset widgets, respectively.
   * More units beside pixels and percentages can be chosen from (the same ones available in GIMP). These units are recalculated automatically.
+* Rename: Removed the `Rename images` option in Export Images due to being redundant there.
 * `Matching text...`: Added `Does not start with text`, `Does not contain text` and `Does not end with text` options.
 * Edit Layers: The `Selected in GIMP` constraint is now added on the first run/after reset.
 * File arguments can now be cleared (set to an empty value). This is useful if a procedure allows omitting a file.